+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## WIP
+- add `Lock` and `LockStake` tx data
+- restore support of `MoveStake` tx data
+- **BREAKING** add fields to VoteCommission tx data
+
 ## 0.41.0 - 2021.12.27
 - add `ReplaceCoinId`, `ReplaceCoinIdByPath`, `GetCoinSymbol` API methods
 - **BREAKING** remove deprecated `getCoinInfoById` API method, use `getCoinInfo` instead